[Silver IV] Title: 수 찾기, Time: 120 ms, Memory: 28648 KB -BaekjoonHub

+++ 백준/Silver/1920. 수 찾기/수 찾기.cs
@@ -7,12 +7,12 @@ StringBuilder sb = new StringBuilder();
 
 int n = int.Parse(sr.ReadLine());
 string[] str = sr.ReadLine().Split();
-Dictionary<BigInteger, int> numDic = new Dictionary<BigInteger, int>(n);
+Dictionary<int, int> numDic = new Dictionary<int, int>(n);
 for(int i = 0; i < n; i++)
-    numDic[BigInteger.Parse(str[i])] = 1;
+    numDic[int.Parse(str[i])] = 1;
 
 sr.ReadLine();
-BigInteger[] biArr = Array.ConvertAll(sr.ReadLine().Split(), BigInteger.Parse);
-foreach(BigInteger a in biArr)
+int[] biArr = Array.ConvertAll(sr.ReadLine().Split(), int.Parse);
+foreach(int a in biArr)
     sb.AppendLine(numDic.ContainsKey(a) ? "1" : "0");
 sw.Write(sb.ToString());
